refactor(admin): change admin-component to edit
