+++ next-tavla/src/Admin/scenarios/AddTile/index.tsx
@@ -1,13 +1,13 @@
 import React, { useState } from 'react'
 import { TTile, TTileType } from 'types/tile'
 import { Button } from '@entur/button'
-import { useSettingsDispatch } from 'Admin/reducer'
 import { RadioGroup, RadioPanel } from '@entur/form'
 import classes from './styles.module.css'
 import { AddStopPlaceTile } from './components/AddStopPlaceTile'
 import { AddQuayTile } from './components/AddQuayTile'
 import { AddMapTile } from './components/AddMapTile'
 import { TAnonTile } from 'Admin/types'
+import { useSettingsDispatch } from 'Admin/utils/contexts'
 
 const components: Record<
     TTileType,
