+++ next-tavla/src/Admin/scenarios/SelectLines/index.tsx
@@ -3,7 +3,7 @@ import { ExpandablePanel } from '@entur/expand'
 import { TQuayTile, TStopPlaceTile } from 'types/tile'
 import { uniqBy } from 'lodash'
 import classes from './styles.module.css'
-import { useSettingsDispatch } from 'Admin/reducer'
+import { useSettingsDispatch } from 'Admin/utils/contexts'
 
 function SelectLines<T extends TStopPlaceTile | TQuayTile>({
     tile,
