+++ next-tavla/src/Admin/scenarios/TilesSettings/index.tsx
@@ -13,7 +13,7 @@ import {
     restrictToVerticalAxis,
 } from '@dnd-kit/modifiers'
 import React from 'react'
-import { useSettingsDispatch } from 'Admin/reducer'
+import { useSettingsDispatch } from 'Admin/utils/contexts'
 import { StopPlaceSettings } from './components/StopPlaceSettings'
 import { QuaySettings } from './components/QuaySettings'
